Create some skeleton files
The first thing I'm going to be working on is the netcdf reader. The
grib reader might not be done for a while, but delay.py is where the
real meat is going to be, since that's the file what will compute the
delay at a particular point, once the weather model has been read in.