Write netcdf.py
I haven't tested it at all yet, and there's much that will change.
Perhaps the entirety of this file will vanish when we move this stuff
onto the GPU. Well, maybe, but it's a good start, and it should have
reasonable performance since all of the hard work is done by scipy.