Refactor code into util.py and fix bugs in delay
I think the whole pipeline works now, except for some confusion about
how I'm going to represent the look vector. The solution for zenith is
really easy, but the difficulty may lie in how to represent it in the
general case. Probably looking at how it comes from the orbit will give
me insight into this.