Deal with NaN in delay calculations
So scipy will use nan outside of the given area when interpolating,
which causes the delay to be calculated as nan a lot of the time. This
is not great, since we want the delay to be 0 above the area for which
we have data, and to repeat the lowest value below. Before we get that
functionality working, though, we can just change nan to 0, because that
gives a delay of 0, which is like there's nothing there. So this should
work for the upper atmosphere, but it's a little bit worse than having
the interpolation work properly.