Write a bad integrator for comparison with scipy
It's only bad in that it's really slow and does a terrible job at
guessing step. I guess those are the only ways that matter. But, I mean,
uh, it gives reasonably accurate results, and with very little work it
could be made to run on the GPU (I'm kidding, it'd be a lot of work).