Fix NaN and partial WV pressure calculation
I'm now using the calculation for the partial pressure of water vapor
from TRAIN. The one from Hobiger et al., 2008 was giving me a negative
answer for the partial pressure, which is of course wrong. So maybe I
typed it in incorrectly, but it's easier just to use TRAIN. And we're
fairly confident that TRAIN is correct, so I think we'll be ok. We just
need to do some garbage that TRAIN does to interpolate below the terrain
and then I'll be fully happy.

Oh yeah and there's the issue where scipy's integrator gives 0 when the
limit of integration is too high. Well, we'll see what we can do about
that.