Add a function to map the delay over a region
Hopefully this is the nice function which will be used mostly by end
users. I've made it parallel and have a sweet progress bar all in the
name of making it easy to use. The only issue now is that it's still
kind of slow.

But, I don't really want to change how it's calculated before our
meeting on Friday. That being said, I could start to implement a second
system, and then only demo this first system on Friday.... That seems
promising. What will the second system be? It must be optimised for
performance.