Use exponentially decaying intervals
When integrating, this particular function I expect to decay
exponentially (the delay function), so it makes sense to do the
intervals of integration like this, rather than have them equally
spaced. I'm about to do a comparison with scipy's built-in integrator,
though, so we'll see how accurate these numbers are.