Make a function for calculating the look vector
I don't know if we'll keep using ecef coordinates forever, but while we
are, it's nice to be able to conveniently convert between lla and ecef,
so I'm making it as easy as possible to do that. I don't love taking the
look vector in as an ecef vector, but it's by far the most
straightforward thing to do.