Make the calculation of e more confusing
So really what I've changed here is that there was an equation from
Hanssen about how to calculate e, so I put in a comment saying that
there's a discrepancy between the two values. When I calculate both,
they sometimes differ by quite a bit, so I'm wondering now what I should
do in order to be the most precise possible.