Use pyproj to convert LLA to ECEF coordinates
I tried this before, but I just didn't know the pyproj order of
arguments. I read it off of stackexchange, and now I'm set to jet as
they say. It gives mostly the same answers as my function, although
sometimes they're off by a little bit. I can only imagine that the new
answer is more accurate, although one can never be sure. Maybe it's
using 32 bit arithmetic instead of 64 and so rounding errors make this
an issue. Maybe not. Hard to say.