Create requirements.txt
It contains all the dependencies for the project at this time. It's good
to stay organized. I don't really want to write code that runs on the
GPU, but it'll go really fast and that will be neat to see. But I won't
get to see it since my laptop doesn't have a GPU. Ah well, what can you
do?