Interpolate delay rather than components directly
It's hard to say what impact this will have on accuracy, but it has a
huge impact on speed. Huge enough that I think this approach should be
used. It also makes it a lot easier to do calculations on the GPU, since
then the GPU is only integrating, and doesn't have to do any complex
calculations. I'm happy with this change, but I'm using git in case I
need to go back.