Propagate actually down
We can check in which order the data come so that we are guaranteed to
propagate in the correct direction. Right now we're assuming pressure is
higher towards the ground and lower away from the ground, but that's
always going to be true, right? Actually, I could use geopotential
heights instead, but that's kind of no better and harder to do.