Deal with range in a more sane way
Now there's not look_vec and range, instead look_vec is just the right
length to begin with. This necessitates the change, though, that if you
pass in Zenith you have to also be able to provide the range. Well,
that's easy, we just put the range into the Zenith class, and there you
go. So that's the solution.