Add a more general purpose delay over region
We've also added a convenience function to read from GDAL files and
calculate the delay over those files. It's the kind of situation where i
make functions when I need the functionality, rather than because it'd
actually be good to have. Fortunately, the public interface is nowhere
near finalized, and so I can change it as much as I want.