Perform interpolation over primitive values
This way we don't have to guess so much about where the error comes in.
However, after performing tests on Mexico, it appears that this change
actually makes almost no difference, with the largest difference between
the two approaches only being 4 mm, which is probably small enough that
I don't care.

On the other hand, though, contrary to what I was saying earlier, this
method isn't actually that much slower than the other. I think my
non-knowledge of numpy was what was slowing us down earlier.