Rework the parallelism
Now it actually goes fast, thanks to some copy-on-write hackery. I don't
love using global variables, since it feels like this could be easily
broken if someone tries to call delay_from_grid in parallel, but also
there's kind of no other option if we want it to run fast. I bet we
could make our own thing to actually use copy-on-write via fork()
directly, but it wouldn't be as easy as using multiprocessing.