Add a function to calculate delay with slant
I have no idea what interface we want for this. That's something that
we'll need to discuss probably in the next meeting. For now, we have a
function which takes in an orbit file and lat, lon, and height
resolutions. Maybe it'd be a good design to move all of the user-facing
code into its own function so that it's clear what may someday change.
But also, it doesn't really matter.