Write a custom interpolator and helper routines
scipy's interpolator has some questionable behavior, so I've written my
own. I've also made some helper routines for raytracing. I know that
these should be in different commits, but that's too much work at this
point. I'm sorry.

Note that this code is really bad in general. I've only committed it
because it's just working and I want to save it. The next commit will
include a fixup and refactoring.