Improve the custom interpolator
It's really slow now, but provides pretty accurate results at least on
one test dataset. We'll first focus on making it even better, then think
about how we can speed it up. We could, of course, do it in parallel,
but it's better to make it quicker single-threaded than just rely on
parallelism for everything.