Clean up lines and make my interpolator default
I think my interpolator is mature enough to be the default. There are
two issues left: it's kind of slow, and it doesn't match TRAIN. Note
that the second issue is present when using the scipy interpolator as
well, and the scipy interpolator is slower to actually process the
image, whereas my interpolator is slower to create the weather model
object.