Redo the interpolation again
Now the interpolation uses only 'griddata', so I hope it's as quick as
possible, and also we're doing cubic interpolation in the vertical
direction, so I hope that's fast. But most importantly, we're really
really close to TRAIN now, so I'm feeling good. Any changes we make from
here on out will have to get better than this excellent commit.