Use multithreading instead of multiprocessing
With numpy we still get a speedup above single-threaded performance,
since numpy can deactivate the GIL. Furthermore, this is faster once we
start processing a lot of data, since with multiprocessing the data has
to be copied and picked and a lot of garbage in order to be sent back.
Here, we just have it in memory. Cool.