Fix some issues in the command-line interface
Yay it's working now!! All that's left is to handle multiple dates and
the automatic downloading of weather model files. I wonder if we
actually want to support multiple dates? The alternative is that the
user just calls the script in a for loop, and that's really not much of
a burden at all.