Fix more small issues
I wish that pylint were easier to use and hated my style less, because
it looks like pyflakes is not going to work for my purposes. Probably
the thing to do is just to fix all the errors that pylint throws, but
that's going to be annoying as it's super super picky.