Remove NaNs from CSV
The program search for intersections between stations and gridcells based off of the user input bounding box. If stations fall outside of the bounding box, no intersection will be found and a NaN is passed to the dataframe. These NaN values were not purged and later crash at the variogram stage. This has been addressed here.