Ignore D212
D212 is "no line break at the beginning of a docstring". Conforming to it fully will involve more than what the automatic fix for this category does (it disregards that the description must also be short enough to fit on one line), so I'm choosing to have ruff ignore it for now. Remind me and I'll make a separate issue for this.