Remove duplicate function test.__init__.py::update_yaml
Two versions of update_yaml existed, the other of which was in tools.RAiDER.aria.prepFromGUNW.py. In order to keep the code clean I extracted the function out to utilFcns.py and redirected imports to point to it instead of the other versions.
This function was also linted to use pathlib.Path.