ruff: ignore ANN101, D200, and D205
ANN101 is adding the type annotation "Self" to every method's self argument. This rule is controversial and in my opinion would be of little benefit for this project.
D200 and D205 are both about the exact formatting of docstrings. I am electing not to make any fixes for these rules since it would involve revising so documentation. This could be revisited in the future with project experts' input.