Type annotate RAiDER run configuration
This change affects the configuration options passed as a YAML file that RAiDER uses to run. Seeing that that is integral to lots of RAiDERs' functionality, it has touched a lot of files.
- Run configuration has been changed from an untyped AttributeDict to a typed dataclass called RunConfig. Right now, checking and validation work is split across several functions in validators.py, and in checkArgs and raider.py, and in smaller places across the program. Storing run configuration in a dataclass may open up possibilities for some of the argument parsing and validating code to be simplified and centralized to the class's constructor.

- Changed --download_only's default from None to False
- Made use of pathlib.Path in a few more places
- Changed some instances of the pattern `if x in [list, of, options]:` to use a tuple instead of a list (runs slightly faster at no cost)
- coerce_into_date now returns a date and not a datetime, as the time portion was unused.
- date_type has been changed to use coerce_into_date under the hood, whereas before it copied its logic and was an instance of data duplication.
- Added comments

I apologize if some of the input validation/parsing is sloppier than it was before; I made just the changes necessary to allow it to work with the dataclass approach. The project is now open for future work where this part of the program can be simplified.