Refactor to allow ruff to identify bare excepts
Replaced `except Exception` and `except BaseException` with just `except`. All are functionally equivalent, but ruff sadly only picks up on bare `except`, so this change was necessary to allow these occurrences to be identified by ruff in the future.
Unfortunately, `except Exception as e` is still not picked up by ruff.